Resolve Enemy Spawning bug
I had issues testing the code I'd written after tying in the networking code, but this patch fixes the sceneID issue. I moved the spawning code into the manager itself and just axed the spawner script entirely
