+++ Assets/Prefabs/EnemyManager.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: dd08b167fd0f39542977c9ffa6b36b98
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
