+++ Assets/Scenes/MenuScene.unity
@@ -4930,14 +4930,6 @@ PrefabInstance:
     serializedVersion: 3
     m_TransformParent: {fileID: 0}
     m_Modifications:
-    - target: {fileID: 866066013436264815, guid: 47c3151aeba853d4ca6323c384aa6dc7, type: 3}
-      propertyPath: spawnPrefabs.Array.size
-      value: 4
-      objectReference: {fileID: 0}
-    - target: {fileID: 866066013436264815, guid: 47c3151aeba853d4ca6323c384aa6dc7, type: 3}
-      propertyPath: spawnPrefabs.Array.data[3]
-      value: 
-      objectReference: {fileID: 6181564224677165571, guid: bb9b1c2b1946df84ead1c59852497d86, type: 3}
     - target: {fileID: 3162481371258705236, guid: 47c3151aeba853d4ca6323c384aa6dc7, type: 3}
       propertyPath: m_LocalPosition.x
       value: 0
