+++ Assets/Scenes/NoProcGen.unity
@@ -1448,7 +1448,7 @@ GameObject:
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 0
+  m_IsActive: 1
 --- !u!114 &504806540
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -3604,56 +3604,6 @@ Transform:
   m_Children: []
   m_Father: {fileID: 1671106592}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!1 &1376223422
-GameObject:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  serializedVersion: 6
-  m_Component:
-  - component: {fileID: 1376223424}
-  - component: {fileID: 1376223423}
-  m_Layer: 0
-  m_Name: EnemyManager
-  m_TagString: Untagged
-  m_Icon: {fileID: 0}
-  m_NavMeshLayer: 0
-  m_StaticEditorFlags: 0
-  m_IsActive: 1
---- !u!114 &1376223423
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1376223422}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 0a696974c6f73e44a85d2e9b7694afb1, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  BuildingsManager: {fileID: 1671106594}
-  Spawner:
-    SquadTemplates:
-    - SquadSO: {fileID: 11400000, guid: 7dbabceac5fa37e45b88e55d94b4eb9a, type: 2}
-      Amount: 3
-  AllEnemies: []
---- !u!4 &1376223424
-Transform:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1376223422}
-  serializedVersion: 2
-  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 0, y: 0, z: 0}
-  m_LocalScale: {x: 1, y: 1, z: 1}
-  m_ConstrainProportionsScale: 0
-  m_Children: []
-  m_Father: {fileID: 0}
-  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &1403988571
 GameObject:
   m_ObjectHideFlags: 0
@@ -5417,7 +5367,6 @@ SceneRoots:
   - {fileID: 36009024}
   - {fileID: 1596879799}
   - {fileID: 1671106592}
-  - {fileID: 1376223424}
   - {fileID: 1018948043}
   - {fileID: 1218696542}
   - {fileID: 1278587104}
